+++ source/SemVerAnalyzer/Assembly/AssemblyPublicInterface.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-
+using System.Runtime.Versioning;
 using dnlib.DotNet;
 
 using Pushpay.SemVerAnalyzer.Engine;
@@ -13,6 +13,7 @@ namespace Pushpay.SemVerAnalyzer.Assembly
 	public class AssemblyPublicInterface
 	{
 		public string Version { get; private set; }
+		public string Framework{ get; private set; }
 		public IReadOnlyList<TypeDef> Types { get; private set; }
 		public IReadOnlyList<AssemblyReference> References { get; private set; }
 
@@ -53,6 +54,8 @@ namespace Pushpay.SemVerAnalyzer.Assembly
 			var informationalVersionAttribute = attributeData.FirstOrDefault(d => d.AttributeType.Name == nameof(AssemblyInformationalVersionAttribute));
 			var informationVersion = informationalVersionAttribute.ConstructorArguments[0].Value.ToString().Split('+')[0];
 			var assemblyReferenceCount = module.TablesStream.AssemblyRefTable.Rows;
+			var targetFrameworkAttribute = attributeData.FirstOrDefault(d => d.AttributeType.Name == nameof(TargetFrameworkAttribute));
+			var targetFramework = targetFrameworkAttribute.ConstructorArguments[0].Value.ToString();
 			var references = new List<AssemblyReference>();
 			// apparently the table is 1-indexed
 			// https://github.com/0xd4d/dnlib/blob/master/src/DotNet/ModuleDefMD.cs#L588
@@ -66,6 +69,7 @@ namespace Pushpay.SemVerAnalyzer.Assembly
 
 			var publicInterface = new AssemblyPublicInterface {
 				Types = module.Types.Where(t => t.IsPublic).ToList(),
+				Framework = targetFramework,
 				Version = SemverExtensions.GetSemVer(informationVersion),
 				References = references
 			};