+++ RELEASE-NOTES.md
@@ -1,6 +1,10 @@
+# v2.3.0
+
+[#52](https://github.com/pushpay-labs/semantic-versioning-analyzer/issues/54) - Automatically detect which target framework to compare.
+
 # v2.2.0
 
-[#52](https://github.com/pushpay-labs/semantic-versioning-analyzer/issues/52)
+[#52](https://github.com/pushpay-labs/semantic-versioning-analyzer/issues/52) - Specify target framework to compare.
 
 # v2.1.1